+++ rsocket/statemachine/StreamsFactory.cpp
@@ -79,7 +79,7 @@ void StreamsFactory::createStreamRequester(
 void StreamsFactory::createStreamRequester(
     Reference<yarpl::flowable::Subscriber<Payload>> responseSink,
     StreamId streamId,
-    uint32_t n) {
+    size_t n) {
   if (connection_.isDisconnectedOrClosed()) {
     subscribeToErrorFlowable(std::move(responseSink));
     return;
